Fix invalid access token error
`fresh_access_token_required` is successfully refreshing the access
token and saving it to the database, but the `station` view page is
reading a stale value from the database.

Until I find a better solution, for now it makes sense to just cache the
access token in the request session. The value will be updated on each
page load.

Fixes #2