Decouple MusicPlayer from view and server
Before this change, MusicPlayer had direct knowledge of the view and
server components so it could initialize them when ready. This was
mostly done because the view and server needed to know when the
MusicPlayer was ready, so they could subscribe to callbacks and know if
they could send play/getVolume/setVolume requests.

Now, MusicPlayer can collect callbacks before Spotify.Player has
initialized and will automatically add them when Spotify.Player is
ready. For the play/pause state, I went with an isReady property on
MusicPlayer that clients can check before calling playback functions.
Another option could be to implement all of the required APIs directly
on MusicPlayer and put the check in there (with a silent ignore).

Given the number of APIs MusicPlayer would now have to implement, I
thought isReady is a more pragmatic approach at this time.