Load StationApp2 from onSpotifyWebPlaybackSDKReady
This change now means that all of the StationApp won't load until
Spotify has loaded. This was required because SystemJS loads the station
module asynchronously and Spotify was ready before the StationApp was
ready. This does simplify the logic in StationApp, as it can assume that
Spotify is available already, so no more need for the stored callbacks
array. However, in the current design, this does mean that the UI logic
can't load until Spotify is ready. At some point it may be worth it to
decouple the UI logic from the Music Player logic - probably when I
actually add the UI logic back in :)