Remove player_state_change event
This creates a feedback loop:
1. LocalPlaybackState changes
2. Updates server state
3. In response, client adjusts local playback state
4. GOTO 1

While yes, if the timing were 100% reliable, the client would not need
to adjust the local playback state in response to the server, in
practice the client is doing micro-adjustments to the playback state,
which is changing the local playback state and firing the event.
Instead, a heartbeat is used to ensure the client/server are in sync.