Remove aiohttp dependency
aiohttp was optimal when Spotify access tokens were being refreshed from
the channels layer, which is asynchronous. Now that access token refresh
has moved to the API layer, which is synchronous, it is no longer
necessary to support asynchronous HTTP requests. This removes one
primary dependency and several transitive dependencies.

Django is inherently synchronous, so supporting asynchronous execution
when not necessary complicates the code and introduces performance
overhead.