Enforce permissions on the Station list and update APIs (#45)
This enforces permissions on listing and updating stations, and creating, listing, and deleting listeners. With this change, if the user is not a listener of the station, it cannot see or update the station. And if the user is not an admin of the station, they cannot create, list, or delete listeners.

This also enforces permissions on creating, listing, and deleting listeners from stations. Wit

Related: #37