Fix station listening (#50)
Replace Django Channels websocketbridge implementation with a [reconnecting-websocket](https://www.npmjs.com/package/reconnecting-websocket). websocketbridge was removed in a recent release of Django Channels, which caused 5xx errors because the static file could not be found.

This also adds a default median ping value in case the ping API times out or can't be reached. Otherwise, this causes all playbackstate change updates to fail because the median time is invalid, which creates an invalid JS Date object.

How Verified:

* can now listen to music and ping information is updating correctly

Fixes #49